dags/requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-fab
pyspark==4.0.0